+++ src/Job/Handler/ReceptionHandler.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Core\Job\Handler;
 
 use Heptacom\HeptaConnect\Core\Job\Contract\ReceptionHandlerInterface;
+use Heptacom\HeptaConnect\Core\Job\Exception\ReceptionJobHandlingException;
 use Heptacom\HeptaConnect\Core\Job\JobData;
 use Heptacom\HeptaConnect\Core\Job\JobDataCollection;
 use Heptacom\HeptaConnect\Core\Job\Type\Reception;
@@ -17,19 +18,21 @@ use Heptacom\HeptaConnect\Portal\Base\Mapping\MappedDatasetEntityStruct;
 use Heptacom\HeptaConnect\Portal\Base\Mapping\TypedMappedDatasetEntityCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\RouteKeyInterface;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\RouteKeyCollection;
 use Heptacom\HeptaConnect\Portal\Base\Support\Contract\DeepObjectIteratorContract;
 use Heptacom\HeptaConnect\Storage\Base\Contract\EntityMapperContract;
 use Heptacom\HeptaConnect\Storage\Base\Contract\EntityReflectorContract;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Repository\JobRepositoryContract;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Repository\MappingNodeRepositoryContract;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Repository\RouteRepositoryContract;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\Get\RouteGetActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\Get\RouteGetCriteria;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\Get\RouteGetResult;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
+use Heptacom\HeptaConnect\Storage\Base\Enum\RouteCapability;
 use Symfony\Component\Lock\LockFactory;
 
 class ReceptionHandler implements ReceptionHandlerInterface
 {
-    private RouteRepositoryContract $routeRepository;
-
     private LockFactory $lockFactory;
 
     private StorageKeyGeneratorContract $storageKeyGenerator;
@@ -46,8 +49,9 @@ class ReceptionHandler implements ReceptionHandlerInterface
 
     private JobRepositoryContract $jobRepository;
 
+    private RouteGetActionInterface $routeGetAction;
+
     public function __construct(
-        RouteRepositoryContract $routeRepository,
         LockFactory $lockFactory,
         StorageKeyGeneratorContract $storageKeyGenerator,
         EntityReflectorContract $entityReflector,
@@ -55,9 +59,9 @@ class ReceptionHandler implements ReceptionHandlerInterface
         MappingNodeRepositoryContract $mappingNodeRepository,
         ReceiveServiceInterface $receiveService,
         DeepObjectIteratorContract $objectIterator,
-        JobRepositoryContract $jobRepository
+        JobRepositoryContract $jobRepository,
+        RouteGetActionInterface $routeGetAction
     ) {
-        $this->routeRepository = $routeRepository;
         $this->lockFactory = $lockFactory;
         $this->storageKeyGenerator = $storageKeyGenerator;
         $this->entityReflector = $entityReflector;
@@ -66,40 +70,56 @@ class ReceptionHandler implements ReceptionHandlerInterface
         $this->receiveService = $receiveService;
         $this->objectIterator = $objectIterator;
         $this->jobRepository = $jobRepository;
+        $this->routeGetAction = $routeGetAction;
     }
 
     public function triggerReception(JobDataCollection $jobs): void
     {
         $receptions = [];
+        $routeKeys = new RouteKeyCollection(\iterable_map(
+            $jobs->column('payload'),
+            static fn (?array $p): ?RouteKeyInterface => $p[Reception::ROUTE_KEY] ?? null
+        ));
+        $routeDatas = $this->routeGetAction->get(new RouteGetCriteria($routeKeys));
+        /** @var RouteGetResult[] $routes */
+        $routes = [];
+
+        foreach ($routeDatas as $routeData) {
+            $routes[$this->storageKeyGenerator->serialize($routeData->getKey())] = $routeData;
+        }
 
         /** @var JobData $job */
         foreach ($jobs as $job) {
             $routeKey = $job->getPayload()[Reception::ROUTE_KEY] ?? null;
 
             if (!$routeKey instanceof RouteKeyInterface) {
-                // TODO error
-                continue;
+                throw new ReceptionJobHandlingException($job, 1636503503);
             }
 
             $entity = $job->getPayload()[Reception::ENTITY] ?? null;
 
             if (!$entity instanceof DatasetEntityContract) {
-                // TODO error
-                continue;
+                throw new ReceptionJobHandlingException($job, 1636503504);
             }
 
-            $route = $this->routeRepository->read($routeKey);
+            $route = $routes[$this->storageKeyGenerator->serialize($routeKey)] ?? null;
+
+            if (!$route instanceof RouteGetResult) {
+                throw new ReceptionJobHandlingException($job, 1636503505);
+            }
+
+            if (!\in_array(RouteCapability::RECEPTION, $route->getCapabilities())) {
+                throw new ReceptionJobHandlingException($job, 1636503506);
+            }
 
             if ($route->getEntityType() !== \get_class($entity)) {
-                // TODO error
-                continue;
+                throw new ReceptionJobHandlingException($job, 1636503507);
             }
 
             $externalId = $job->getMappingComponent()->getExternalId();
 
             if (!\is_string($externalId)) {
-                // TODO error
-                continue;
+                throw new ReceptionJobHandlingException($job, 1636503508);
             }
 
             $targetPortal = $this->storageKeyGenerator->serialize($route->getTargetKey());